Remove inner classes
